prevent deployments of Next.js apps vulnerable to CVE-2025-66478 (#9572)
* prevent deployments of Next.js apps vulnerable to CVE-2025-66478

* Update CHANGELOG to reference CVE-2025-66478 prevention in Next.js deployments (#9572)

* Refactor error message for Next.js CVE-2025-66478 vulnerability check

Updated the error handling to provide a more detailed and version-specific message when a vulnerable Next.js version is detected. The message now specifies the required patched versions based on the major version of Next.js, enhancing clarity for users on how to resolve the issue.

* Enhance Next.js vulnerability check documentation

Added references to CVE-2025-66478 and CVE-2025-55182 in the documentation for the Next.js version vulnerability check function. This provides users with direct links to relevant security advisories, improving awareness and guidance on addressing potential vulnerabilities.

* format

* update next.js version in webframeworks-deploy-tests

* use semver for error message checks

* remove unnecessary comments

* Revert "update next.js version in webframeworks-deploy-tests"

This reverts commit 22857f430094761c0f1f8ab183b12d4ddd133d1d.

---------

Co-authored-by: Joe Hanley <joehanley@google.com>